refactor: remove BlockEditor and BlockItem components, integrate TiptapBlockEditor with useDocumentBlocks hook
