+++ client/src/components/Textarea.tsx
@@ -1,29 +1,25 @@
-import { useAutoResize } from "@/hooks";
 import React, { useCallback, useMemo } from "react";
 
 interface Props {
   value: string;
   onChange: (value: string) => void;
-  isTitleDocument?: boolean;
   onFocus?: () => void;
   onKeyDown?: (e: React.KeyboardEvent) => void;
   onBlur?: () => void;
   className?: string;
+  placeholder?: string;
 }
 
 export const Textarea = React.memo(
   ({
-    isTitleDocument,
     onFocus,
     onKeyDown,
     onBlur,
     value,
     onChange,
     className,
+    placeholder,
   }: Props) => {
-    const textareaRef = useAutoResize(value);
-
-    // Memoize change handler
     const handleChange = useCallback(
       (e: React.ChangeEvent<HTMLTextAreaElement>) => {
         onChange(e.target.value);
@@ -31,30 +27,16 @@ export const Textarea = React.memo(
       [onChange]
     );
 
-    // Memoize styles and classes
-    const textareaStyles = useMemo(
-      () => ({
-        minHeight: "auto",
-      }),
-      []
-    );
-
     const textareaClasses = useMemo(
       () =>
-        `w-full text-4xl font-bold bg-transparent border-none outline-none resize-none placeholder-gray-400 ${
+        `w-full bg-transparent border-none outline-none resize-none placeholder-gray-400 ${
           className || ""
         }`.trim(),
       [className]
     );
 
-    const placeholder = useMemo(
-      () => (isTitleDocument ? "Untitled" : "Type '/' for commands"),
-      [isTitleDocument]
-    );
-
     return (
       <textarea
-        ref={textareaRef}
         onKeyDown={onKeyDown}
         onFocus={onFocus}
         value={value}
@@ -62,7 +44,6 @@ export const Textarea = React.memo(
         onBlur={onBlur}
         className={textareaClasses}
         placeholder={placeholder}
-        style={textareaStyles}
       />
     );
   }
