+++ client/src/hooks/index.ts
@@ -2,9 +2,9 @@ export { useAuth, useUserId, useIsAuthenticated, useToken } from './use-auth';
 export { useLogout } from './use-logout';
 export { useBlocks } from './use-blocks';
 export { useWorkspace } from './use-workspace';
-export { useAutoResize } from './use-auto-resize';
 export { useLoading, useMultipleLoading } from './use-loading';
 export { useToast, toast } from './use-toast';
 export { useBinCraftTitle } from './use-bin-craft-title';
 export { useDebounce } from './use-debounce';
-export type { Block, CreateBlockInput } from './use-blocks'; 
+export type { Block, CreateBlockInput } from './use-blocks';
+export { useDocumentBlocks } from './use-document-blocks';
